MODLD-745: STORY: API to assign a preferred instance profile to a user (#215)
* MODLD-745: STORY: API to assign a preferred instance profile to a user

* MODLD-745: Add tests

* MODLD-745: Update NEWS.md file

* MODLD-745: Refactor ProfileServiceImpl to simplify code

* MODLD-745: Change the type of profile ID from Long to Integer

+++ src/main/java/org/folio/linked/data/service/tenant/worker/ProfileWorker.java
@@ -52,5 +52,5 @@ public class ProfileWorker implements TenantServiceWorker {
     }
   }
 
-  record Profile(Long id, String name, String resourceType, JsonNode value) {}
+  record Profile(Integer id, String name, String resourceType, JsonNode value) {}
 }
